Divide up gcode emulator and loader
