+++ meerk40t/grbl/gcodeloader.py
@@ -1,0 +1,18 @@
+import os
+
+
+class GCodeLoader:
+    @staticmethod
+    def load_types():
+        yield "Gcode File", ("gcode", "nc", "gc"), "application/x-gcode"
+
+    @staticmethod
+    def load(kernel, service, pathname, **kwargs):
+        basename = os.path.basename(pathname)
+        with open(pathname, "rb") as f:
+            op_branch = service.get(type="branch ops")
+            op_branch.add(
+                data=list(f.readlines()), data_type="grbl", type="blob", name=basename
+            )
+            kernel.root.close(basename)
+            return True
